refactoring render and remove
