+++ src/presenter/film-presenter.js
@@ -1,4 +1,4 @@
-import { render } from '../render.js';
+import { render, remove } from '../framework/render.js';
 import { isEscapeEvent } from '../util.js';
 import FilmCardView from '../view/film-card-view.js';
 import PopupView from '../view/popup-view.js';
@@ -38,8 +38,7 @@ export default class FilmPresenter {
   };
 
   #closePopup = () => {
-    this.#popupComponent.element.remove();
-    this.#popupComponent.removeElement();
+    remove(this.#popupComponent);
     document.body.classList.remove('hide-overflow');
     document.removeEventListener('keydown', this.#onEscKeyDown);
   };
